[score boosting] support datetime expressions (#6196)
* use f64 as internal expression score

* allow datetime as expression

* add test cases

* fix schema, generate openapi and grpc docs

* fmt

* clippy

* fix test precision

* graceful handling of too large numbers, more tests

* rename to DatetimeExpression for consistency

* use datetime's timestamp as seconds

* capture variable name when parsing

* homogenize DateTime into Datetime

* use interface distinction between datetime strings and payload keys

* fix rebase

* fix after rebase

+++ lib/common/common/src/math.rs
@@ -17,13 +17,13 @@ pub fn scaled_fast_sigmoid(x: ScoreType) -> ScoreType {
     0.5 * (fast_sigmoid(x) + 1.0)
 }
 
-pub fn is_close(a: f32, b: f32) -> bool {
-    const ABS_TOL: f32 = 1e-6;
-    const REL_TOL: f32 = 1e-6;
+pub fn is_close(a: f64, b: f64) -> bool {
+    const ABS_TOL: f64 = 1e-6;
+    const REL_TOL: f64 = 1e-6;
     is_close_tol(a, b, ABS_TOL, REL_TOL)
 }
 
-pub fn is_close_tol(a: f32, b: f32, abs_tol: f32, rel_tol: f32) -> bool {
+pub fn is_close_tol(a: f64, b: f64, abs_tol: f64, rel_tol: f64) -> bool {
     let tol = a.abs().max(b.abs()) * rel_tol + abs_tol;
     (a - b).abs() <= tol
 }
